requirements.txt
requests
python-dotenv
anthropicresend
resend
google-search-results